refactor: rename the 'vo' pathname e classes to 'valueobjects'
